requirements.txt
streamlit>=1.35.0
anthropic>=0.25.0
chromadb>=0.5.0
pypdf>=4.2.0
sentence-transformers>=2.7.0
python-docx>=1.1.0
python-dotenv>=1.0.0
pandas>=2.0.0
